add support for f128 parsing
This does not implement the eisel-lemire algorithm for f128, but does
implement the fast and slow-paths. The existing code-path for f64 or
smaller uses the same types as internal types have been specialized
specifically for f128. This ensures performance for the more widely used
floating point types is not impacted by f128 support.

The slow-path decimal shift table-code has been reworked to use a more
direct set of constants. The previous implementation (based on rust
dec2flt) encoded the data in a non-obvious and was a pain to update.
There is no practical performance difference between the two.

All f128 tests pass when using the parse-number-fxx-test-data repo.